requirements.txt
# AI Object Detection Studio v6.1 - Requirements
# Cài đặt: pip install -r requirements.txt

# Core AI & Computer Vision
ultralytics>=8.0.0          # YOLOv8 framework
torch>=2.0.0                # PyTorch deep learning
torchvision>=0.15.0         # PyTorch vision utilities
opencv-python>=4.8.0        # Computer vision library

# Image Processing
Pillow>=10.0.0              # Python Imaging Library
numpy>=1.24.0               # Numerical computing

# GUI Framework
customtkinter>=5.2.0        # Modern tkinter interface

# Data Handling & Export
pandas>=2.0.0               # Data manipulation
matplotlib>=3.7.0           # Plotting and visualization

# Development & Monitoring
tensorboard>=2.13.0         # Training visualization (optional)

# Optional: Enhanced functionality
# tkinterdnd2>=0.3.0        # Drag & drop support (uncomment if needed)
# pygame>=2.5.0             # Audio notifications (uncomment if needed)
# requests>=2.31.0          # HTTP requests for cloud features (uncomment if needed)

# System Requirements (automatically installed)
setuptools>=68.0.0
wheel>=0.41.0